using math/big types for confirmation calculation

+++ packages/go-kosu/store/witness.go
@@ -5,5 +5,5 @@ import "go-kosu/abci/types"
 // TransactionWitness extends the TransactionWitness with Confirmations
 type TransactionWitness struct {
 	types.TransactionWitness
-	Confirmations uint32
+	Confirmations uint64
 }
